server/requirements.txt
alembic==1.6.5
amqp==5.0.6
appdirs==1.4.4
attrs==21.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cliff==3.8.0
cmd2==2.0.1
colorama==0.4.4
cryptography==3.4.7
debtcollector==2.2.0
decorator==5.0.9
dnspython==1.16.0
dogpile.cache==1.1.3
eventlet==0.31.0
extras==1.0.0
fasteners==0.16.2
fixtures==3.0.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
future==0.18.2
futurist==2.3.0
greenlet==1.0.0
httplib2==0.19.1
idna==2.10
iso8601==0.1.14
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
jsonpatch==1.32
jsonpointer==2.1
jsonschema==3.2.0
# Editable Git install with no remote (junco==1.0)
# -e /home/comunica/DEV/junco/server
junos-eznc==2.5.4
keystoneauth1==4.3.1
keystonemiddleware==9.3.0
kombu==5.1.0
linecache2==1.0.0
logutils==0.3.5
lxml==4.6.3
Mako==1.1.4
MarkupSafe==1.1.1
msgpack==1.0.2
munch==2.5.0
ncclient==0.6.10
netaddr==0.8.0
netifaces==0.11.0
netmiko==3.3.3
networking-generic-switch==5.0.0
neutron==18.0.0
neutron-lib==2.12.0
nslookup==1.4.0
ntc-templates==2.0.0
numpy==1.20.2
openstacksdk==0.57.0
os-client-config==2.1.0
os-ken==2.0.0
os-service-types==1.7.0
os-traits==2.5.0
os-vif==2.5.0
osc-lib==2.4.0
oslo.cache==2.8.0
oslo.concurrency==4.4.0
oslo.config==8.7.0
oslo.context==3.3.0
oslo.db==9.0.0
oslo.i18n==5.0.1
oslo.log==4.5.0
oslo.messaging==12.8.0
oslo.middleware==4.2.0
oslo.policy==3.8.0
oslo.privsep==2.5.0
oslo.reports==2.2.0
oslo.rootwrap==6.2.0
oslo.serialization==4.1.0
oslo.service==2.5.0
oslo.upgradecheck==1.4.0
oslo.utils==4.9.0
oslo.versionedobjects==2.4.0
osprofiler==3.4.0
ovs==2.13.3
ovsdbapp==1.10.0
packaging==20.9
pandas==1.2.3
paramiko==2.7.2
Paste==3.5.0
PasteDeploy==2.1.1
pbr==5.6.0
pecan==1.4.0
prettytable==0.7.2
psutil==5.8.0
pycadf==3.1.1
pycparser==2.20
PyMySQL==1.0.2
PyNaCl==1.4.0
# Pyodbc module
pyodbc==4.0.32

pyOpenSSL==20.0.1
pyparsing==2.4.7
pyperclip==1.8.2
pyreadline3==3.3
pyrsistent==0.17.3
pyserial==3.5
python-dateutil==2.8.1
python-designateclient==4.3.0
python-dotenv==0.19.0
python-dracclient==6.0.0
python-editor==1.0.4
python-keystoneclient==4.2.0
python-mimeparse==1.6.0
python-neutronclient==7.4.0
python-novaclient==17.5.0
pytz==2021.1
PyYAML==5.4.1
repoze.lru==0.7
requests==2.25.1
requestsexceptions==1.4.0
rfc3986==1.5.0
Routes==2.5.1
scp==0.13.3
setproctitle==1.2.2
simplejson==3.17.2
six==1.15.0
sortedcontainers==2.4.0
soupsieve==2.2.1
SQLAlchemy==1.4.6
sqlalchemy-migrate==0.13.0
sqlparse==0.4.1
statsd==3.3.0
stevedore==3.3.0
tacacs-plus==2.6
Tempita==0.5.2
tenacity==6.3.1
testresources==2.0.1
testscenarios==0.5.0
testtools==2.4.0
textfsm==1.1.0
tooz==2.9.0
traceback2==1.4.0
transitions==0.8.8
unittest2==1.1.0
urllib3==1.26.5
vine==5.0.0
voluptuous==0.12.1
waitress==2.0.0
wcwidth==0.2.5
WebOb==1.8.7
WebTest==2.0.35
Werkzeug==1.0.1
wrapt==1.12.1
yamlordereddictloader==0.4.0
yappi==1.3.2
